+++ packages/kit/test/apps/basics/svelte.config.js
@@ -30,7 +30,13 @@ const config = {
 				'/routing/prerendered/trailing-slash/never',
 				'/routing/prerendered/trailing-slash/ignore'
 			],
-			handleHttpError: 'warn'
+			handleHttpError: ({ path, message }) => {
+				if (path.includes('/reroute/async')) {
+					throw new Error('shouldnt error on ' + path);
+				}
+
+				console.warn(message);
+			}
 		},
 
 		version: {